Storage.java: Abstract out the copyPreviousFiles method into several lower-level methods
Within the copyPreviousFiles lie the functionality to read different lines of the saved text file. Depending on what the kind of task it is (ToDo, Deadline, Event), the method provides different ways of processing the line.

Abstracting the processes into lower-level methods serves to trim the code and provide better readability.

Further, this method also processes tasks based on whether they have been marked as done. This function has been abstracted into a separate, lower-level method to improve code readability.

Due to the method's implementation of streams and AtomicIntegers, the previous code was cluttered and disorganized. This has been improved upon.