Update creation of Tasks from .txt file
Storage.java handles the reading and creating of the Task objects, when loading from janet.txt file in directory. Storage.java also needs to check whether the content of the .txt file is valid.

Shifting the creation of the Task objects to the Task themselves ensure that the validity checks are also performed in the Task objects. This makes it easier to perform validity checks through exception handling.