Deadline.java: Add check when date is unexpected
Deadline object's dueDate can be earlier than today's date.

A check ensures that the dueDate specified by the user is not earlier than today's date.

Let's add a check in the creation of the Deadline object from user's input to ensure that the dueDate cannot be earlier than today's date.