Event.java: Add check when date is unexpected
Event object's startDate can be earlier than today's date. The startDate can also be later than or equal to endDate

A check ensures that the startDate specified by the user is not earlier than today's date. The startDate is also not later than or equal to endDate.

Let's add a check in the creation of the Event object from user's input to ensure that the startDate cannot be earlier than today's date, and startDate cannot be later than or equal to endDate.