Add assert statement
The parseInput method returns an Action that has a Command attribute.

The executeAction method receives the Action and executes the respective actions based on the value of the Command attribute.

The Command attribute may have an invalid value due to mistakes in the code.

Let’s add an assert statement that checks if the value of the command attribute of the action returned by parseInput is a valid command.

This allows the runtime to detect a possible bug.