Update README.md
README does not contain the information
about the chat bot.

Users would not be able to know what the
chat bot do when reading the README.

Let's update the README with images
and information about the features
of the chat bot.

Updating the README can allow users
to understand how to use the chat bot
and the features available before opening
the application.