Add method to fetch file
File path is hardcoded.

File to save and load tasklist will not be created
if file does not exist causing file not found error to
occur whenever the chat bot is started.

Let's create a method, fetchFile() to create a file
if the file is not present in the specific file path.
The method is called before loading the storage

Creating a file when the file does not exist will
not result in file not found exception and allows
user to continue saving task into tasklist.
Method is called before loading storage so that
there is for sure a file to load and save the
data into.