Improve code quality
Name of functions in tasklist are not informative.

More descriptive function names make it easier to understand the code.

Let's change the names to be more descriptive.

UI is an argument for multiple functions despite not being used.

This makes the code less readable and may confuse users expecting UI to be used.

Let's remove the UI from the function arguments.