Add Ui.png and User Guide
Lack of relevant User guide and picture of messenger interface.

The product is less intriging to potential users due to the lack
of visual aids and a consumer-friendly explanation of what the
product does.

Let's add a .png file showcasing the first thing the user sees
when they run the product. A proper User Guide also helps users
to better understand how to use / navigate the product!