Modify execute() method
There are multiple levels of abstraction found within a code fragment
in the execute() method, which is related to the deletion of a task
from the task list and storage.

Extracting the relevant code and creating multiple functions to handle
different aspects of the deletion function will allow for greater
readability of the code, making it easier to understand and debug,
if necessary.