Modify execute() function
There are multiple levels of abstraction found within a code fragment
in the execute() method, which is related to finding tasks with
descriptions containing the input match string from user.

Extracting the relevant code and creating multiple functions to handle
different aspects of the matching function will allow for greater
readability of the code, making it easier to understand and debug,
if necessary.