Add deleteall functionality, separate Parser into different classes, check for duplicates when adding Tasks, and add support for aliases when entering dates (e.g. today, tomorrow, etc.).
The Parser currently handles all user input and error handling for user
input. This is violation of SLAP.

Separating Parsers for different types of parsing helps reinforce OOP
principles and keep better track of error handling across different
across different errors encountered.

Let's extract a CommandParser to make sense of which Commands to be
executed by the system, a DateTimeParser to parse dates and times,
an IndexParser to make sense of indexes entered, and a TaskParser to
make sense of information entered into Task objects.

Following SLAP is important for code readability, quality, and ease for
debugging when encountering unexpected errors.