TaskList.java: Handle duplicated task description
TaskList.java allows tasks with the same descriptions to be added into the list of tasks.

Not allowing tasks with the same descriptions to be added makes it more user friendly because tasks should belong only to either a ToDo, Deadline or Event.

Let's add a check in the addTaskToList method in TaskList.java to prevent the addition of new tasks with descriptions that are identical to at least one other Task in the current list of tasks.