Alter database path for JAR
Database creation / access path is based off the layout in IntelliJ

The many different folders and directory's in the layout of Intelllij
mis different from that when the user downloads the JAR file into their
own folder, making it vulnerable to errors when run as a JAR file in a
folder that does not have the same layout.

Let's change the path to make it not dependent on other folders, and
just have it created in the same directory as the JAR file.