Modify javadoc for TaskList::findTasks method
The current javadoc description for the
method does not fully explain what instance
fields the query matches when the method
checks for a match.

Changing the javadoc to specify that the
description of the task is used for matching
better explains the purse of the method.

Let's modify the javadoc to specify that
string matching is done with the task
description to improve the documentation
for the function.