Rainy.java: Improve update functionality
This increment is decidedly important as it gives users the ability
to edit their previous task additions.

Since there are three different types of tasks, I had to separately
implement the functionality for each different task.

For ToDo, it is the most simple since it only takes in a string, it
accepts any kind of text input.

For Deadline, users are required to specify whether they want to
change the name of the task or the date of the task.

Similarly, for event, users need to specify which parameter they are
changing, the name, date, or time of the event.

All other functionalities of the chatbot remain intact.