Change behaviour of fromString method
The current implementation attempts to match
the given input with the appriopriate CommandType
enum value. If the input does not match any
enum values, an IllegalArgumentException is thrown.

This behaviour uses an exception for control flow,
which should be avoided.

Adding a special type to signify invalid command
types might be a better way to handle bad inputs.

Let's
- Add an INVALID type for CommandType
- Use the INVALID type as the default return value
of CommandType::fromString