Refactor bot's return values after commands
The current implementation for getResponse uses
stores information about the previous executed
command as a workaround for javafx components to
access information about the executed command.

This behaviour contributes to the temporary field
code smell, as additional null checks need to be
made if public methods are not called in the right
order.

Extracting out such information and placing them in
a record class might be better to move data related
to commands.

Lets
- Create a record object Response to store the data
need by JavaFX for rendering
- Refactor command objects to return a string response
in the execute method instead of storing it as an instance
variable to be accessed later
- Clean up other parts of the program that require
said data to use the Response record instead.
- Remove unneeded instance fields that clutter the
classes mentioned