Fix bug by including isDone in Fixed constructor's super call
The Fixed constructor does not reference the isDone
parameter when calling the superclass constructor, causing fixed
tasks to always be read as unmarked.

This needs to change to correctly reflect the completion status of
fixed tasks.

Include the isDone parameter in the call to the superclass
constructor.

It ensures that the fixed tasks correctly maintain and display their
completion status.