Refactor SecondMind and TaskList classes
There are multiple levels of abstraction found within code fragments
in SecondMind and TaskList classes, which are related to the cretaion
of different Task Objects.

Extracting the relevant code and creating multiple functions to handle
different aspects of the task creation and adding tasks to the TaskList
will allow for greater readability of the code, making it easier to
understand and debug, if necessary.