Remove dead methods
There is one dead initialisation method in CommandAlternatives class.

This may be confusing for new developers to the project because they
may not understand the reason for the dead method, what purpose it
serves.

Let's remove the dead method from the project.

This will improve code quality and readability for new developers to
the project.