Setup data for GUI screenshot
The current storage data is not suitable for data displaying.

This needs to change to ensure a reasonable data is being output for the
screenshot purpose in future.

Edit the data in the data/lama.txt to make the data more reasonable.

This is done to ensure the data is better for displaying purpose.

No changes to functionality.