remove .swp files from tracking
The .swp files generated by the editor are meant to be ignored as they
are temporary files.

Previously, the .Orion.java.swp file was tracked by Git. This commit
removes it from tracking while keeping it locally to avoid cluttering
the repository with temporary files.

Let's ensure that .swp files will not be tracked again by adding them
to the .gitignore file. This improves the repository's cleanliness and
reduces unnecessary changes being tracked.