Add assert to getTaskLists with invalid task type
Since the save file is managed by the program itself, there should never be
an instance where there is a task type that is not recognised by the program
itself. If such a case occur, it means that the save file have
been tampered with and data integrity has been breached.