Improve UI and add new help command
Add the following changes to UI:
- new House Stark background
- colourings for dialog boxes depending if it's user or ned
- red colouring for dialog box if it's an error message
- dynamic resizing of window
- prompt text to user input
- change send button message to 'decree'
Add new help command, which is activated by user inputting 'help'.

The UI was generic and lacking of a theme suited to the personality of
its chatbot personality.
A help command to guide users on which commands and how to use them is
not present.