Fix code quality in the Gale class
Certain methods in the Gale class could be further abstracted out.
The methods are: findTasks, handleTaskMarking and deleteTask.

Let's create new helper methods to abstract out the logic within
the aforementioned methods and place them in their respective
classes.

In the Parser class, the parseIndex and parseKeyword methods were
created.

In the TaskList class, the markTask and deleteTask methods were
edited.