use assert features
assert is used to make sure the task names cannot be null in the getName methods.

assert is used to make sure there is a history and task list object made for Bwead.

assert is used to make sure the task list contains more than one tasks when the delete command is used.

assert is used to make sure the task list is not empty, especially after adding a new task to it.