Improve GUI
MoiMoi's GUI displays both sides of the conversation in the same visual
format.

This needs to change to better reflect the asymmetric nature of
interactions between the user and the program.

Let's tweak the GUI design to distinguish between user and MoiMoi dialog
boxes.

This makes the user interface more visually intuitive, and improves user
experience.