Add FindTest class for testing FindCommand
The FindTest class is important as it cannot be tested
in a vacuum due to its nature, yet is an important
part of our implementation.

To this end, the FindTest class provides the smallest
abstraction of the Majimabot that does not involve the
GUI but is more concerned with the intended behaviour
of the FindCommand.