Update userImage and babyImage field in MainWindow
The two images are declared as field read from a certain
path, but in case the path does not exist or image files
are not working, the loadImage() method complements with a
transparent image.

This is applied so that it is promised the software works
whether given the presence of the images.