Add assert to document assumptions
Switch cases within the code makes assumption that the correct input is
parsed and given.

There are thus no default case within the switches since we assume that
all inputs will correctly map to one of the possible case.

Using assert function will ensure that this assumption is correct and
will help in testing the code in fringe cases.

The asserts are deemed as true and placed as the default cases as the
assumption we are making is that the code will never reach those default
cases of those switches.

Other edge cases within the code will throw an error and be handled
within the program itself and thus assert should not be used for those
cases.