Adjust the size of the profile images for both user and gopher
The current profile images are unnecessarily big, because the users and
the chatbot are fixed characters in the application, this results in a
lot of valuable spaces in the window wasted.

Hence, the profile images are reduced in size to give more space to the
text and other UI elements.