Implement save and load API for taskList Object
Currently save and load task are handled by the Task Manager Object,
which requires reference of the ArrayList of tasks inside the TaskList
Object to work correctly. Such implementation is very inflexible as
TaskList object is only accessible within the Gopher class but its
ArrayList of task cannot be accessed due to encapsulation.

Therefore, save and load task API are implemented within TaskList class
as an application of the Tell, Don't Ask Principle to ensure that Gopher
class can get access to save and load feature for its own task list
without having to break the abstraction barrier.