Fix the issue of tasks statuses not updated when InvalidTaskNumberException is thrown.
Currently, task statuses will not get updated correctly because of
premature return of the functions due to InvalidTaskNumberException,
this makes it very hard for user to keep track of the actual status
of the task list, hence rendering the system useless.

Therefore, finally blocks are added to execute the taskList save API
to ensure that task statuses are always saved regardless of whether
an exception is thrown during the program execution.