Fix the issue of Gopher congratulating user when no matched task found
Currently, gopher will display a message congratulating user for not
having pending task when there's no task that appear on the search
result. This message is misleading and should not be there when find
method is executed.

Hence, this message has been extract to UI class and would only appear
when list task method is called and not when the find task method is
called. This ensures that the gopher's message to the user is always
accurate