Add new edit feature to tasks
If a user wants to edit a specific part of a task, they would need to
delete the original task and then add a completely new task.

Let's add a new feature that can directly edit a part of an existing
task without having to use multiple commands. For instance, in an event
task, the edit command can allow just the end time to be updated.

Having an edit feature is preferable as it is easier to make
direct changes to tasks that still need to be on the task list.