Refactor code for better code quality
Certain methods within the class had different levels of abstraction and
did not follow SLAP. Within the commands that had to manipulate the
inputs, data processing was done together with other high level
abstraction. Loading save file error is not reflected in the GUI.

The code needs to be easily read and understood so that modifications
in the future can be easily done. Errors are also more prone to happen
to code that are not easily understandable.

Shifting the data processing within the command classes to its own
separate function allows for the code to follow SLAP and be more
succinct.

Data processing of the inputs should still be done within the command
class as it is consistent with the expected behaviour thus separating
data processing to a separate function is preferred over processing the
data in another class.