Improve error handling
Parsing an integer in "mark", "unmark" and "delete" command does
not account for the case where the input is of an invalid format.

Handling this error can improve user experience as they will get
a feedback when this error occurs.

Let's catch this error and print it out using the Ui.