Add multiple Command classes
Currently the run() method in the parser class does not adhere to
the SLAP principle. There are various levels of abstractions in the
run() method.

Multiple Command classes that inherits the Command interface are
added so that the logic are implemented in their respective Command
class. This improves code quality.