Split commands into multiple command classes
lemon.executeLemon(): very long method that needs to be refactored

The extensiveness of the function can be hard to troubleshoot or add new commands. Additionally the error handling of the function is not done very clearly.

By refactoring and removing duplicated code can help clean up the code and improve code quality.

Change:
* parser class: Handle the switch cases for generating the right command to be executed
* command classes: Handle the implementation of the commands and necessary exception handling
* ui class: Handle printing of error messages
* Remove InvalidCommandException and DescriptionException classes as they are no longer used

Todo:
* More extensive testing of lemon.executeLemon() to make sure the new version does not have unexpected errors
* Add the necessary JavaDocs for newly implemented classes and methods once finalized