Polish GUI and chat bot appearance
The GUI is plain and lacks a sense of feedback for user input.

The resulting user experience is unsatisfying and uninspired as the
chat bot appears generic compared to other chat bots.

Let's
* Add colour chat bot's replies and vary the colour of the dialog box
according to the type of command given.
* Add colour to the text input field.
* Add a background image to separate this chat bot from others and make
the background image dark so that the dialog box stands out against the
background.
* Add a placeholder text in the text input field to more clearly
indicate to the user where the type their input.
* Add a title and name for the chat bot app.