Extend InvalidTokenException check to task creation feature
Currently, user can create task with certain token names if these tokens
are not recognized by the respective task creation command. This may
lead to confusion of user when viewing the task list content.

Therefore, the InvalidTokenException check is extended to all types of
task creation command to ensure that user will be warned that they can't
use tokens for other tasks as part of the task name, hence improving the
overall user experience.