Change add tasks methods
There were two add tasks methods to handle the difference in format of
the tasks that were stored from the previous use vs the tasks entered
into the command line by the user.

Since these two methods have the same logic, let's merge them together
so we don't have duplicate code.

Modifying the format in which tasks are stored so that they can be parsed
easily, and modifying the way in which tasks are parsed to allow alternate
formats.