Fix the issue of task list delete feature removing the wrong task
Currently the delete feature in task list will remove the first n tasks
if user provides n task numbers, which is not correct. It is found that
the retrieving of task from the list is not working properly because the
tasks are accessed using loop control variable i instead of the supplied
task numbers.

Hence, the delete function is changed to ensure that it accesses the
correct tasks based on the given task numbers instead.