Add JUnit tests for Tayoo#getResponse
There is no test for getResponse() in Tayoo to check that it sets the
field for isExit and commandType. If future commits alter the behaviour
of getResponse() or its dependent methods, or create new commandTypes
with incorrect return values for these variables, it could cause the
chat bot to malfunction.

Let's create a test to check that the getResponse() function works as expected.

That way, if any future changes to the code base alters either the
commands or getResponse, the JUnit tests will highlight it out to the
developer.