Fix find function
Find function returns entire list of tasks instead of correctly
filtered tasks.

This is a bug and will confuse the user and may make them stop using
the app.

Let's fix the bug so the find functionality can be used properly.

This will allow the app to be used seamlessly again.